made iso_defaults public and static
This is really just 6324 without deleting Phreeqc.h and keeping the history. In order
to keep the history you need to do a reverse merge.

see http://stackoverflow.com/questions/2863028/recover-a-folder-or-file-in-tortoisesvn-whilst-also-retaining-all-history

answer attached:

Do a "reverse merge":
 1.Ensure that your working copy is updated to HEAD and completely clean (not strictly necessary, but always a good idea before trying to merge)
 2.Right click on the folder and select TortoiseSVN > Merge...
 3.Select "Merge a range of revisions" and hit "Next"
 4.The URL to merge from is the repo path for your current directory (ie. the one containing the deleted file)
 5.The revision range to merge is the revision in which you deleted the file (show log might help here)
 6.Be sure to select "Reverse merge"
 7.Hit "Next"
 8.All the default merge options should be sufficient, so hit "Merge"
 
The file should now be added and still retain all of its previous history.

Explanation:

This "rolls back" the revision containing the deletion, but adds the merge information to the directory properties. SVN can track the original file through this metadata. On the command line you would to do a backwards merge of the revision in which you deleted it:
svn merge -c -<revision-number> path/containing/file

Note the hyphen before the rev number (ie. a "negative" revision nunber).
 

git-svn-id: svn://136.177.114.72/svn_GW/phreeqc3/trunk@6328 1feff8c3-07ed-0310-ac33-dd36852eb9cd